==> R/process/00_reprocess_vswc_data.R <==
# Download and reprocess soil water content sensor data using consistent calibration coefficients for the entire time series
# Clear the environment to avoid using info from previous site

# Adapted from Ed Ayres -
# https://figshare.com/articles/software/Code_for_SMAP_NLDAS_validations_based_on_NEON_soil_moisture_data/24424762/1?file=43528533


# Create loop to run through a few sites at a time
information <- tibble(site = c("SRER","SJER","WREF","UNDE","WOOD","KONZ"),
                      start = c("2022-05","2022-05","2022-05","2024-05","2024-05","2024-05"),
                      end = c("2022-06","2022-06","2022-06","2024-06","2024-06","2024-06")
)

swc_results <- vector(mode="list",length = nrow(information))


reprocess_vswc <- function(site,start,end) {

  # Replace data with NA to prevent any carryover to next loop
  swc <- NA
  SWS_30_minute <- NA
  sensor_positions_00094 <- NA
  neonBoulded60Cal <- NA
  neonCal <- NA

  # download the data
  swc <- neonUtilities::loadByProduct(dpID="DP1.00094.001", site=site, startdate=start, enddate=end, package="expanded", timeIndex="30", check.size=F, nCores=3,include.provisional = TRUE)
  list2env(swc, .GlobalEnv)
  SWS_30_minute <- swc$SWS_30_minute
  # Change dates to posixct
  SWS_30_minute$startDateTime <- as.POSIXct(SWS_30_minute$startDateTime, tz="GMT")
  startYYYYMM <- substr(SWS_30_minute$startDateTime[1], 1, 7)

  # # save metadata files
  # positionsFilename <- paste0("N:/Science/FIU/NASA_SMAP/Reprocessed SWC Sensor Data/positions_", site, "_", startYYYYMM, "_to_", end, ".csv")
  # write.csv(sensor_positions_00094, file =positionsFilename)
  # readmeFilename <- paste0("N:/Science/FIU/NASA_SMAP/Reprocessed SWC Sensor Data/readme_", site, "_", startYYYYMM, "_to_", end, ".csv")
  # write.csv(readme_00094, file =readmeFilename)
  # variablesFilename <- paste0("N:/Science/FIU/NASA_SMAP/Reprocessed SWC Sensor Data/variables_", site, "_", startYYYYMM, "_to_", end, ".csv")
  # write.csv(variables_00094, file =variablesFilename)

  # Read in NEON calibrations
  neonCal <- read.csv("data/raw/SoilMoistureDepths_FromAs-built_DontOpenInExcel.csv", header=T, stringsAsFactors=F)
  # Read in NEON bounded60 calibrations
  neonBoulded60Cal <- read.csv("data/raw/Bounded60_SoilMoistureDepths_FromAs-built_DontOpenInExcel.csv", header=T, stringsAsFactors=F)


  # Identify rows calculated with the Sentek manufacturer calibration based on calDefaultQM and VSWCExpUncert
  # defaultCalRows <- intersect(which(SWS_30_minute$calDefaultQM > 0), which(SWS_30_minute$VSWCExpUncert > 0.21)) # filtering by which(SWS_30_minute$VSWCExpUncert > 0.21) was causing rows to be missed in recent parts of time series (after ~2020-2021)
  defaultCalRows <- which(SWS_30_minute$calDefaultQM > 0)
  # Add rows where defaults calibration was used (uncertainty > 0.21), but calibration coefficients were not reported in the cal file so the calibration flag did not work correctly
  startDateCalFlagP1 <- SWS_30_minute$startDateTime[intersect(which(SWS_30_minute$horizontalPosition == "001"), defaultCalRows)][1]
  startDateCalFlagP2 <- SWS_30_minute$startDateTime[intersect(which(SWS_30_minute$horizontalPosition == "002"), defaultCalRows)][1]
  startDateCalFlagP3 <- SWS_30_minute$startDateTime[intersect(which(SWS_30_minute$horizontalPosition == "003"), defaultCalRows)][1]
  startDateCalFlagP4 <- SWS_30_minute$startDateTime[intersect(which(SWS_30_minute$horizontalPosition == "004"), defaultCalRows)][1]
  startDateCalFlagP5 <- SWS_30_minute$startDateTime[intersect(which(SWS_30_minute$horizontalPosition == "005"), defaultCalRows)][1]
  defaultCalRows <- append(defaultCalRows, intersect(intersect(which(SWS_30_minute$startDateTime < startDateCalFlagP1), which(SWS_30_minute$horizontalPosition == "001")), which(SWS_30_minute$VSWCExpUncert > 0.21)))
  defaultCalRows <- append(defaultCalRows, intersect(intersect(which(SWS_30_minute$startDateTime < startDateCalFlagP2), which(SWS_30_minute$horizontalPosition == "002")), which(SWS_30_minute$VSWCExpUncert > 0.21)))
  defaultCalRows <- append(defaultCalRows, intersect(intersect(which(SWS_30_minute$startDateTime < startDateCalFlagP3), which(SWS_30_minute$horizontalPosition == "003")), which(SWS_30_minute$VSWCExpUncert > 0.21)))
  defaultCalRows <- append(defaultCalRows, intersect(intersect(which(SWS_30_minute$startDateTime < startDateCalFlagP4), which(SWS_30_minute$horizontalPosition == "004")), which(SWS_30_minute$VSWCExpUncert > 0.21)))
  defaultCalRows <- append(defaultCalRows, intersect(intersect(which(SWS_30_minute$startDateTime < startDateCalFlagP5), which(SWS_30_minute$horizontalPosition == "005")), which(SWS_30_minute$VSWCExpUncert > 0.21)))


  # Calculate sensor scaled frequency for the default calibration rows (based on Sentek Calibration Manual). Multiply water content by 100 to get % vol units
  SWS_30_minute$sf <- NA
  SWS_30_minute$sf[defaultCalRows] <- (0.1957 * ((SWS_30_minute$VSWCMean[defaultCalRows]*100)^0.404)) + 0.02852


  # Identify rows calculated with the NEON calibration based on calDefaultQM and VSWCExpUncert
  neonCalRowsP1D1 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "501"))
  neonCalRowsP2D1 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "501"))
  neonCalRowsP3D1 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "501"))
  neonCalRowsP4D1 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "501"))
  neonCalRowsP5D1 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "501"))
  ################################# Measurement level 2
  neonCalRowsP1D2 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "502"))
  neonCalRowsP2D2 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "502"))
  neonCalRowsP3D2 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "502"))
  neonCalRowsP4D2 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "502"))
  neonCalRowsP5D2 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "502"))
  ################################# Measurement level 3
  neonCalRowsP1D3 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "503"))
  neonCalRowsP2D3 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "503"))
  neonCalRowsP3D3 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "503"))
  neonCalRowsP4D3 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "503"))
  neonCalRowsP5D3 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "503"))
  ################################# Measurement level 4
  neonCalRowsP1D4 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "504"))
  neonCalRowsP2D4 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "504"))
  neonCalRowsP3D4 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "504"))
  neonCalRowsP4D4 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "504"))
  neonCalRowsP5D4 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "504"))
  ################################# Measurement level 5
  neonCalRowsP1D5 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "505"))
  neonCalRowsP2D5 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "505"))
  neonCalRowsP3D5 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "505"))
  neonCalRowsP4D5 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "505"))
  neonCalRowsP5D5 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "505"))
  ################################# Measurement level 6
  neonCalRowsP1D6 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "506"))
  neonCalRowsP2D6 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "506"))
  neonCalRowsP3D6 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "506"))
  neonCalRowsP4D6 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "506"))
  neonCalRowsP5D6 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "506"))
  ################################# Measurement level 7
  neonCalRowsP1D7 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "507"))
  neonCalRowsP2D7 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "507"))
  neonCalRowsP3D7 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "507"))
  neonCalRowsP4D7 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "507"))
  neonCalRowsP5D7 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "507"))
  ################################# Measurement level 8
  neonCalRowsP1D8 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "001"))
                               , which(SWS_30_minute$verticalPosition == "508"))
  neonCalRowsP2D8 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "002"))
                               , which(SWS_30_minute$verticalPosition == "508"))
  neonCalRowsP3D8 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "003"))
                               , which(SWS_30_minute$verticalPosition == "508"))
  neonCalRowsP4D8 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "004"))
                               , which(SWS_30_minute$verticalPosition == "508"))
  neonCalRowsP5D8 <- intersect(intersect(intersect(which(SWS_30_minute$calNEONQM > 0), which(SWS_30_minute$VSWCExpUncert < 0.21))
                                         , which(SWS_30_minute$horizontalPosition == "005"))
                               , which(SWS_30_minute$verticalPosition == "508"))
  # Create vector of plots and measurement levels where soil moisture will be recalculated
  neonCals <- ls()[grep("neonCalRowsP", ls())]


  # Calculate sensor scaled frequency for the neon calibration rows (based on neonCal data frame)
  for(i in 1:length(neonCals)){
    # Find the row in the calibration data frame corresponding to the sensor
    row <- intersect(intersect(grep(site, neonCal$site), which(neonCal$plot == substr(neonCals[i], 13, 13))), which(neonCal$measurementLevel == substr(neonCals[i], 15, 15)) )
    # Assign the calibration coefficients
    a <- neonCal$CVALA1[row]
    b <- neonCal$CVALA2[row]
    c <- neonCal$CVALA3[row]
    # Identify the rows where the calibration coefficients will be applied
    if(neonCals[i] == "neonCalRowsP1D1"){
      rows <- neonCalRowsP1D1
    }else if(neonCals[i] == "neonCalRowsP2D1"){
      rows <- neonCalRowsP2D1
    }else if(neonCals[i] == "neonCalRowsP3D1"){
      rows <- neonCalRowsP3D1
    }else if(neonCals[i] == "neonCalRowsP4D1"){
      rows <- neonCalRowsP4D1
    }else if(neonCals[i] == "neonCalRowsP5D1"){
      rows <- neonCalRowsP5D1
    }
    #### Measurement level 2
    if(neonCals[i] == "neonCalRowsP1D2"){
      rows <- neonCalRowsP1D2
    }else if(neonCals[i] == "neonCalRowsP2D2"){
      rows <- neonCalRowsP2D2
    }else if(neonCals[i] == "neonCalRowsP3D2"){
      rows <- neonCalRowsP3D2
    }else if(neonCals[i] == "neonCalRowsP4D2"){
      rows <- neonCalRowsP4D2
    }else if(neonCals[i] == "neonCalRowsP5D2"){
      rows <- neonCalRowsP5D2
    }
    #### Measurement level 3
    if(neonCals[i] == "neonCalRowsP1D3"){
      rows <- neonCalRowsP1D3
    }else if(neonCals[i] == "neonCalRowsP2D3"){
      rows <- neonCalRowsP2D3
    }else if(neonCals[i] == "neonCalRowsP3D3"){
      rows <- neonCalRowsP3D3
    }else if(neonCals[i] == "neonCalRowsP4D3"){
      rows <- neonCalRowsP4D3
    }else if(neonCals[i] == "neonCalRowsP5D3"){
      rows <- neonCalRowsP5D3
    }
    #### Measurement level 4
    if(neonCals[i] == "neonCalRowsP1D4"){
      rows <- neonCalRowsP1D4
    }else if(neonCals[i] == "neonCalRowsP2D4"){
      rows <- neonCalRowsP2D4
    }else if(neonCals[i] == "neonCalRowsP3D4"){
      rows <- neonCalRowsP3D4
    }else if(neonCals[i] == "neonCalRowsP4D4"){
      rows <- neonCalRowsP4D4
    }else if(neonCals[i] == "neonCalRowsP5D4"){
      rows <- neonCalRowsP5D4
    }
    #### Measurement level 5
    if(neonCals[i] == "neonCalRowsP1D5"){
      rows <- neonCalRowsP1D5
    }else if(neonCals[i] == "neonCalRowsP2D5"){
      rows <- neonCalRowsP2D5
    }else if(neonCals[i] == "neonCalRowsP3D5"){
      rows <- neonCalRowsP3D5
    }else if(neonCals[i] == "neonCalRowsP4D5"){
      rows <- neonCalRowsP4D5
    }else if(neonCals[i] == "neonCalRowsP5D5"){
      rows <- neonCalRowsP5D5
    }
    #### Measurement level 6
    if(neonCals[i] == "neonCalRowsP1D6"){
      rows <- neonCalRowsP1D6
    }else if(neonCals[i] == "neonCalRowsP2D6"){
      rows <- neonCalRowsP2D6
    }else if(neonCals[i] == "neonCalRowsP3D6"){
      rows <- neonCalRowsP3D6
    }else if(neonCals[i] == "neonCalRowsP4D6"){
      rows <- neonCalRowsP4D6
    }else if(neonCals[i] == "neonCalRowsP5D6"){
      rows <- neonCalRowsP5D6
    }
    #### Measurement level 7
    if(neonCals[i] == "neonCalRowsP1D7"){
      rows <- neonCalRowsP1D7
    }else if(neonCals[i] == "neonCalRowsP2D7"){
      rows <- neonCalRowsP2D7
    }else if(neonCals[i] == "neonCalRowsP3D7"){
      rows <- neonCalRowsP3D7
    }else if(neonCals[i] == "neonCalRowsP4D7"){
      rows <- neonCalRowsP4D7
    }else if(neonCals[i] == "neonCalRowsP5D7"){
      rows <- neonCalRowsP5D7
    }
    #### Measurement level 8
    if(neonCals[i] == "neonCalRowsP1D8"){
      rows <- neonCalRowsP1D8
    }else if(neonCals[i] == "neonCalRowsP2D8"){
      rows <- neonCalRowsP2D8
    }else if(neonCals[i] == "neonCalRowsP3D8"){
      rows <- neonCalRowsP3D8
    }else if(neonCals[i] == "neonCalRowsP4D8"){
      rows <- neonCalRowsP4D8
    }else if(neonCals[i] == "neonCalRowsP5D8"){
      rows <- neonCalRowsP5D8
    }

    # Back-calculate scaled frequency using the calibration coefficients
    SWS_30_minute$sf[rows] <- (a * ((SWS_30_minute$VSWCMean[rows]*100)^b)) + c

    # prevent carryover to next loop
    row <- NA
    rows <- NA
  }


  # Calculate soil water content (cm3/cm3) using the neon bounded60 calibration coefficients
  SWS_30_minute$correctedVSWCMean <- NA
  for(i in 1:length(neonCals)){
    # Find the row in the calibration data frame corresponding to the sensor
    row <- intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == substr(neonCals[i], 13, 13))), which(neonBoulded60Cal$measurementLevel == substr(neonCals[i], 15, 15)) )
    # Assign the calibration coefficients
    aBounded <- neonBoulded60Cal$CVALA1[row]
    bBounded <- neonBoulded60Cal$CVALA2[row]
    cBounded <- neonBoulded60Cal$CVALA3[row]
    # Identify the rows where the calibration coefficients will be applied
    if(neonCals[i] == "neonCalRowsP1D1"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "501"))
    }else if(neonCals[i] == "neonCalRowsP2D1"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "501"))
    }else if(neonCals[i] == "neonCalRowsP3D1"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "501"))
    }else if(neonCals[i] == "neonCalRowsP4D1"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "501"))
    }else if(neonCals[i] == "neonCalRowsP5D1"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "501"))
    }
    #### Measurement level 2
    if(neonCals[i] == "neonCalRowsP1D2"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "502"))
    }else if(neonCals[i] == "neonCalRowsP2D2"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "502"))
    }else if(neonCals[i] == "neonCalRowsP3D2"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "502"))
    }else if(neonCals[i] == "neonCalRowsP4D2"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "502"))
    }else if(neonCals[i] == "neonCalRowsP5D2"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "502"))
    }
    #### Measurement level 3
    if(neonCals[i] == "neonCalRowsP1D3"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "503"))
    }else if(neonCals[i] == "neonCalRowsP2D3"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "503"))
    }else if(neonCals[i] == "neonCalRowsP3D3"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "503"))
    }else if(neonCals[i] == "neonCalRowsP4D3"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "503"))
    }else if(neonCals[i] == "neonCalRowsP5D3"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "503"))
    }
    #### Measurement level 4
    if(neonCals[i] == "neonCalRowsP1D4"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "504"))
    }else if(neonCals[i] == "neonCalRowsP2D4"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "504"))
    }else if(neonCals[i] == "neonCalRowsP3D4"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "504"))
    }else if(neonCals[i] == "neonCalRowsP4D4"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "504"))
    }else if(neonCals[i] == "neonCalRowsP5D4"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "504"))
    }
    #### Measurement level 5
    if(neonCals[i] == "neonCalRowsP1D5"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "505"))
    }else if(neonCals[i] == "neonCalRowsP2D5"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "505"))
    }else if(neonCals[i] == "neonCalRowsP3D5"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "505"))
    }else if(neonCals[i] == "neonCalRowsP4D5"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "505"))
    }else if(neonCals[i] == "neonCalRowsP5D5"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "505"))
    }
    #### Measurement level 6
    if(neonCals[i] == "neonCalRowsP1D6"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "506"))
    }else if(neonCals[i] == "neonCalRowsP2D6"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "506"))
    }else if(neonCals[i] == "neonCalRowsP3D6"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "506"))
    }else if(neonCals[i] == "neonCalRowsP4D6"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "506"))
    }else if(neonCals[i] == "neonCalRowsP5D6"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "506"))
    }
    #### Measurement level 7
    if(neonCals[i] == "neonCalRowsP1D7"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "507"))
    }else if(neonCals[i] == "neonCalRowsP2D7"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "507"))
    }else if(neonCals[i] == "neonCalRowsP3D7"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "507"))
    }else if(neonCals[i] == "neonCalRowsP4D7"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "507"))
    }else if(neonCals[i] == "neonCalRowsP5D7"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "507"))
    }
    #### Measurement level 8
    if(neonCals[i] == "neonCalRowsP1D8"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "508"))
    }else if(neonCals[i] == "neonCalRowsP2D8"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "508"))
    }else if(neonCals[i] == "neonCalRowsP3D8"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "508"))
    }else if(neonCals[i] == "neonCalRowsP4D8"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "508"))
    }else if(neonCals[i] == "neonCalRowsP5D8"){
      rows <- intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "508"))
    }

    # Apply the calibration
    SWS_30_minute$correctedVSWCMean[rows] <- (((SWS_30_minute$sf[rows] - cBounded) / aBounded)^(1/bBounded)) / 100

    # prevent carryover to next loop
    row <- NA
    rows <- NA
  }

  # Assign corrected water content final quality flag based on alpha and beta quality metrics, temp test, and science review flags
  # Use temperature test to inform flagging at all sites for measurement levels 1-3, but for deeper measurement levels only use temperature test at sites where >=35 cm soil temperature falls below 0.1 degrees C (based on 2016-April 2022 RELEASE 2023 soil temperature data)
  coldSites <- c("BARR", "HEAL", "TREE", "TOOL", "DEJU", "NOGP", "NIWO", "WOOD", "BONA", "DCFS", "CPER", "YELL", "RMNP", "MOAB", "STER", "BLAN", "ONAQ")
  if(length(grep(site, coldSites)) == 1){
    # Identify rows with higher alpha metric, but that pass all other tests (beta, temp test, science review)
    goodRows <- intersect(intersect(intersect(intersect(which(SWS_30_minute$VSWCAlphaQM < 30),
                                                        which(SWS_30_minute$VSWCBetaQM < 20)),
                                              c(which(SWS_30_minute$VSWCFinalQFSciRvw == 0),
                                                which(is.na(SWS_30_minute$VSWCFinalQFSciRvw), arr.ind=TRUE))),
                                    which(SWS_30_minute$tempPassQM > 0)),
                          which(SWS_30_minute$tempFailQM == 0))
  }else{
    goodRowsDeep <- intersect(intersect(intersect(which(SWS_30_minute$VSWCAlphaQM < 30),
                                                  which(SWS_30_minute$VSWCBetaQM < 20)),
                                        c(which(SWS_30_minute$VSWCFinalQFSciRvw == 0),
                                          which(is.na(SWS_30_minute$VSWCFinalQFSciRvw), arr.ind=T))),
                              which(SWS_30_minute$verticalPosition >= 504))
    goodRowsShallow <- intersect(intersect(intersect(intersect(intersect(which(SWS_30_minute$VSWCAlphaQM < 30),
                                                                         which(SWS_30_minute$VSWCBetaQM < 20)),
                                                               c(which(SWS_30_minute$VSWCFinalQFSciRvw == 0),
                                                                 which(is.na(SWS_30_minute$VSWCFinalQFSciRvw), arr.ind=T))),
                                                     which(SWS_30_minute$verticalPosition < 504)),
                                           which(SWS_30_minute$tempPassQM > 0)),
                                 which(SWS_30_minute$tempFailQM == 0))
    goodRows <- c(goodRowsShallow, goodRowsDeep)
  }
  SWS_30_minute$correctedVSWCFinalQF <- 1
  SWS_30_minute$correctedVSWCFinalQF[goodRows] <- 0





  # Assign measurement uncertainty to unflagged rows. NOTE this does not include changes in water content within the averaging period
  SWS_30_minute$correctedVSWCExpUncert <- NA
  goodP1D1 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "501")), goodRows)
  goodP2D1 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "501")), goodRows)
  goodP3D1 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "501")), goodRows)
  goodP4D1 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "501")), goodRows)
  goodP5D1 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "501")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D1] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 1) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D1] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 1) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D1] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 1) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D1] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 1) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D1] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 1) )] *2
  #### Measurement level 2
  goodP1D2 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "502")), goodRows)
  goodP2D2 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "502")), goodRows)
  goodP3D2 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "502")), goodRows)
  goodP4D2 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "502")), goodRows)
  goodP5D2 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "502")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D2] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 2) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D2] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 2) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D2] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 2) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D2] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 2) )] *2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D2] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 2) )] *2
  #### Measurement level 3
  goodP1D3 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "503")), goodRows)
  goodP2D3 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "503")), goodRows)
  goodP3D3 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "503")), goodRows)
  goodP4D3 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "503")), goodRows)
  goodP5D3 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "503")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D3] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 3) )] *2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D3] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 3) )] *2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D3] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 3) )] *2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D3] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 3) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D3] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 3) )] *2
  #### Measurement level 4
  goodP1D4 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "504")), goodRows)
  goodP2D4 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "504")), goodRows)
  goodP3D4 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "504")), goodRows)
  goodP4D4 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "504")), goodRows)
  goodP5D4 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "504")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D4] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 4) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D4] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 4) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D4] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 4) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D4] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 4) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D4] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 4) )]*2
  #### Measurement level 5
  goodP1D5 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "505")), goodRows)
  goodP2D5 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "505")), goodRows)
  goodP3D5 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "505")), goodRows)
  goodP4D5 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "505")), goodRows)
  goodP5D5 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "505")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D5] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 5) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D5] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 5) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D5] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 5) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D5] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 5) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D5] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 5) )]*2
  #### Measurement level 6
  goodP1D6 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "506")), goodRows)
  goodP2D6 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "506")), goodRows)
  goodP3D6 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "506")), goodRows)
  goodP4D6 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "506")), goodRows)
  goodP5D6 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "506")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D6] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 6) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D6] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 6) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D6] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 6) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D6] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 6) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D6] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 6) )]*2
  #### Measurement level 7
  goodP1D7 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "507")), goodRows)
  goodP2D7 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "507")), goodRows)
  goodP3D7 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "507")), goodRows)
  goodP4D7 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "507")), goodRows)
  goodP5D7 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "507")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D7] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 7) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D7] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 7) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D7] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 7) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D7] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 7) )]*2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D7] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 7) )]*2
  #### Measurement level 8
  goodP1D8 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "001"), which(SWS_30_minute$verticalPosition == "508")), goodRows)
  goodP2D8 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "002"), which(SWS_30_minute$verticalPosition == "508")), goodRows)
  goodP3D8 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "003"), which(SWS_30_minute$verticalPosition == "508")), goodRows)
  goodP4D8 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "004"), which(SWS_30_minute$verticalPosition == "508")), goodRows)
  goodP5D8 <- intersect(intersect(which(SWS_30_minute$horizontalPosition == "005"), which(SWS_30_minute$verticalPosition == "508")), goodRows)
  SWS_30_minute$correctedVSWCExpUncert[goodP1D8] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 1)), which(neonBoulded60Cal$measurementLevel == 8) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP2D8] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 2)), which(neonBoulded60Cal$measurementLevel == 8) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP3D8] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 3)), which(neonBoulded60Cal$measurementLevel == 8) )] *2
  SWS_30_minute$correctedVSWCExpUncert[goodP4D8] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 4)), which(neonBoulded60Cal$measurementLevel == 8) )] * 2
  SWS_30_minute$correctedVSWCExpUncert[goodP5D8] <- neonBoulded60Cal$U_CVALA3[intersect(intersect(grep(site, neonBoulded60Cal$site), which(neonBoulded60Cal$plot == 5)), which(neonBoulded60Cal$measurementLevel == 8) )] * 2






  # Plot original L1 data
  minDate <- min(SWS_30_minute$startDateTime)
  maxDate <- max(SWS_30_minute$startDateTime)
  filenameUncorrected <- paste0("figures/Reprocessed SWC Sensor Data/UncorrectedSWC_", site, "_", startYYYYMM, "_to_", end, ".png")
  png(filenameUncorrected, units = "px", width = 1500, height = 1500, res=150)
  main <- paste0("Uncorrected soil moisture ", site, " ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D1], SWS_30_minute$VSWCMean[goodP1D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  points(SWS_30_minute$startDateTime[goodP2D1], SWS_30_minute$VSWCMean[goodP2D1], pch=".", col="red")
  points(SWS_30_minute$startDateTime[goodP3D1], SWS_30_minute$VSWCMean[goodP3D1], pch=".", col="orange")
  points(SWS_30_minute$startDateTime[goodP4D1], SWS_30_minute$VSWCMean[goodP4D1], pch=".", col="green")
  points(SWS_30_minute$startDateTime[goodP5D1], SWS_30_minute$VSWCMean[goodP5D1], pch=".", col="blue")
  legend("topleft", legend=c("Plot 1", "Plot 2", "Plot 3", "Plot 4", "Plot 5"), lty=1, col=c("black", "red", "orange", "green", "blue"), bty="n")
  dev.off()


  # plot scaled frequency
  filenameSF <- paste0("figures/Reprocessed SWC Sensor Data/scaledFreq_", site, "_", startYYYYMM, "_to_", end, ".png")
  png(filenameSF, units = "px", width = 1500, height = 1500, res=150)
  main <- paste0("Corrected scaled frequency ", site, " ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D1], SWS_30_minute$sf[goodP1D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Scaled frequency (proportion)", main=main)
  points(SWS_30_minute$startDateTime[goodP2D1], SWS_30_minute$sf[goodP2D1], pch=".", col="red")
  points(SWS_30_minute$startDateTime[goodP3D1], SWS_30_minute$sf[goodP3D1], pch=".", col="orange")
  points(SWS_30_minute$startDateTime[goodP4D1], SWS_30_minute$sf[goodP4D1], pch=".", col="green")
  points(SWS_30_minute$startDateTime[goodP5D1], SWS_30_minute$sf[goodP5D1], pch=".", col="blue")
  legend("topleft", legend=c("Plot 1", "Plot 2", "Plot 3", "Plot 4", "Plot 5"), lty=1, col=c("black", "red", "orange", "green", "blue"), bty="n")
  dev.off()


  # Plot corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 1500, height = 1500, res=150)
  main <- paste0("Corrected soil moisture ", site, " ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D1], SWS_30_minute$correctedVSWCMean[goodP1D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  points(SWS_30_minute$startDateTime[goodP2D1], SWS_30_minute$correctedVSWCMean[goodP2D1], pch=".", col="red")
  points(SWS_30_minute$startDateTime[goodP3D1], SWS_30_minute$correctedVSWCMean[goodP3D1], pch=".", col="orange")
  points(SWS_30_minute$startDateTime[goodP4D1], SWS_30_minute$correctedVSWCMean[goodP4D1], pch=".", col="green")
  points(SWS_30_minute$startDateTime[goodP5D1], SWS_30_minute$correctedVSWCMean[goodP5D1], pch=".", col="blue")
  legend("topleft", legend=c("Plot 1", "Plot 2", "Plot 3", "Plot 4", "Plot 5"), lty=1, col=c("black", "red", "orange", "green", "blue"), bty="n")
  dev.off()


  # Create graph of soil plot 1 corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_P1_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 2500, height = 1500, res=150)
  par(mfrow = c(2, 4))
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 1 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D1], SWS_30_minute$correctedVSWCMean[goodP1D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 2 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D2], SWS_30_minute$correctedVSWCMean[goodP1D2], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 3 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D3], SWS_30_minute$correctedVSWCMean[goodP1D3], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 4 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D4], SWS_30_minute$correctedVSWCMean[goodP1D4], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 5 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D5], SWS_30_minute$correctedVSWCMean[goodP1D5], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 6 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D6], SWS_30_minute$correctedVSWCMean[goodP1D6], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 7 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D7], SWS_30_minute$correctedVSWCMean[goodP1D7], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 1 Depth 8 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP1D8], SWS_30_minute$correctedVSWCMean[goodP1D8], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  # legend("topleft", legend=c("Level 1", "Level 2", "Level 3", "Level 4", "Level 5", "Level 6", "Level 7", "Level 8"), lty=1, col=c("black", "red", "orange", "yellow", "green", "turquoise", "blue", "purple"), bty="n")
  dev.off()


  # Create graph of soil plot 2 corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_P2_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 2500, height = 1500, res=150)
  par(mfrow = c(2, 4))
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 1 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D1], SWS_30_minute$correctedVSWCMean[goodP2D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 2 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D2], SWS_30_minute$correctedVSWCMean[goodP2D2], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 3 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D3], SWS_30_minute$correctedVSWCMean[goodP2D3], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 4 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D4], SWS_30_minute$correctedVSWCMean[goodP2D4], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 5 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D5], SWS_30_minute$correctedVSWCMean[goodP2D5], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 6 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D6], SWS_30_minute$correctedVSWCMean[goodP2D6], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 7 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D7], SWS_30_minute$correctedVSWCMean[goodP2D7], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 2 Depth 8 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP2D8], SWS_30_minute$correctedVSWCMean[goodP2D8], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  # legend("topleft", legend=c("Level 1", "Level 2", "Level 3", "Level 4", "Level 5", "Level 6", "Level 7", "Level 8"), lty=1, col=c("black", "red", "orange", "yellow", "green", "turquoise", "blue", "purple"), bty="n")
  dev.off()


  # Create graph of soil plot 3 corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_P3_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 2500, height = 1500, res=150)
  par(mfrow = c(2, 4))
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 1 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D1], SWS_30_minute$correctedVSWCMean[goodP3D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 2 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D2], SWS_30_minute$correctedVSWCMean[goodP3D2], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 3 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D3], SWS_30_minute$correctedVSWCMean[goodP3D3], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 4 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D4], SWS_30_minute$correctedVSWCMean[goodP3D4], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 5 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D5], SWS_30_minute$correctedVSWCMean[goodP3D5], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 6 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D6], SWS_30_minute$correctedVSWCMean[goodP3D6], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 7 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D7], SWS_30_minute$correctedVSWCMean[goodP3D7], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 3 Depth 8 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP3D8], SWS_30_minute$correctedVSWCMean[goodP3D8], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  # legend("topleft", legend=c("Level 1", "Level 2", "Level 3", "Level 4", "Level 5", "Level 6", "Level 7", "Level 8"), lty=1, col=c("black", "red", "orange", "yellow", "green", "turquoise", "blue", "purple"), bty="n")
  dev.off()

  # Create graph of soil plot 4 corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_P4_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 2500, height = 1500, res=150)
  par(mfrow = c(2, 4))
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 1 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D1], SWS_30_minute$correctedVSWCMean[goodP4D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 2 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D2], SWS_30_minute$correctedVSWCMean[goodP4D2], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 3 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D3], SWS_30_minute$correctedVSWCMean[goodP4D3], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 4 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D4], SWS_30_minute$correctedVSWCMean[goodP4D4], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 5 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D5], SWS_30_minute$correctedVSWCMean[goodP4D5], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 6 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D6], SWS_30_minute$correctedVSWCMean[goodP4D6], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 7 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D7], SWS_30_minute$correctedVSWCMean[goodP4D7], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 4 Depth 8 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP4D8], SWS_30_minute$correctedVSWCMean[goodP4D8], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  # legend("topleft", legend=c("Level 1", "Level 2", "Level 3", "Level 4", "Level 5", "Level 6", "Level 7", "Level 8"), lty=1, col=c("black", "red", "orange", "yellow", "green", "turquoise", "blue", "purple"), bty="n")
  dev.off()

  # Create graph of soil plot 5 corrected L1 data
  filenameCorrected <- paste0("figures/Reprocessed SWC Sensor Data/correctedSWC_", site, "_P5_", startYYYYMM, "_to_", end, ".png")
  png(filenameCorrected, units = "px", width = 2500, height = 1500, res=150)
  par(mfrow = c(2, 4))
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 1 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D1], SWS_30_minute$correctedVSWCMean[goodP5D1], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 2 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D2], SWS_30_minute$correctedVSWCMean[goodP5D2], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 3 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D3], SWS_30_minute$correctedVSWCMean[goodP5D3], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 4 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D4], SWS_30_minute$correctedVSWCMean[goodP5D4], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 5 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D5], SWS_30_minute$correctedVSWCMean[goodP5D5], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 6 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D6], SWS_30_minute$correctedVSWCMean[goodP5D6], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 7 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D7], SWS_30_minute$correctedVSWCMean[goodP5D7], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  main <- paste0("Corrected SWC ", site, " Plot 5 Depth 8 ", startYYYYMM, " to ", end)
  plot(SWS_30_minute$startDateTime[goodP5D8], SWS_30_minute$correctedVSWCMean[goodP5D8], xlim=c(minDate,maxDate), pch=".", ylim=c(0,1), xlab="", ylab="Soil moisture (cm3/cm3)", main=main)
  # legend("topleft", legend=c("Level 1", "Level 2", "Level 3", "Level 4", "Level 5", "Level 6", "Level 7", "Level 8"), lty=1, col=c("black", "red", "orange", "yellow", "green", "turquoise", "blue", "purple"), bty="n")
  dev.off()

  # save the corrected data in a separate file for each soil plot
  filenameP1 <- paste0("data/derived/Reprocessed SWC Sensor Data/corrected_", site, "_P1_", startYYYYMM, "_to_", end, ".csv")
  write.csv(SWS_30_minute[which(SWS_30_minute$horizontalPosition=="001"), ], row.names = F, file =filenameP1)

  filenameP2 <- paste0("data/derived/Reprocessed SWC Sensor Data/corrected_", site, "_P2_", startYYYYMM, "_to_", end, ".csv")
  write.csv(SWS_30_minute[which(SWS_30_minute$horizontalPosition=="002"), ], row.names = F, file =filenameP2)

  filenameP3 <- paste0("data/derived/Reprocessed SWC Sensor Data/corrected_", site, "_P3_", startYYYYMM, "_to_", end, ".csv")
  write.csv(SWS_30_minute[which(SWS_30_minute$horizontalPosition=="003"), ], row.names = F, file =filenameP3)

  filenameP4 <- paste0("data/derived/Reprocessed SWC Sensor Data/corrected_", site, "_P4_", startYYYYMM, "_to_", end, ".csv")
  write.csv(SWS_30_minute[which(SWS_30_minute$horizontalPosition=="004"), ], row.names = F, file =filenameP4)

  filenameP5 <- paste0("data/derived/Reprocessed SWC Sensor Data/corrected_", site, "_P5_", startYYYYMM, "_to_", end, ".csv")
  write.csv(SWS_30_minute[which(SWS_30_minute$horizontalPosition=="005"), ], row.names = F, file =filenameP5)

  return(SWS_30_minute)
}

for(i in 1:nrow(information)) {
  swc_results[[i]] <- reprocess_vswc(information$site[[i]],information$start[[i]],information$end[[i]])
}

vswc_information <- information |>
  mutate(vswc_corrected = swc_results)

save(vswc_information,file = 'data/derived/reprocessed_vswc_data.Rda')
